Move TFTP and HTTP to their own packages.
