+++ http/http_test.go
@@ -1,4 +1,4 @@
-package ipxe
+package http
 
 import (
 	"bytes"
